feat(assets): add dithered `Alexander visits Diogenes in Corinth` pictures
ALEXANDER DIOGENEM CORINTH VISIT.

"Thereupon many statesmen and philosophers came to Alexander with their congratulations,
and he expected that Diogenes of Sinope also, who was tarrying in Corinth, would do likewise.
But since that philosopher took not the slightest notice of Alexander,
and continued to enjoy his leisure in the suburb Craneion, Alexander went in person to see him,
and he found him lying in the sun. Diogenes raised himself up a little when he saw
so many people coming towards him, and fixed his eyes upon Alexander.
And when that monarch addressed him with greetings, and asked if he wanted anything,
"Yes," said Diogenes, "stand a little out of my sun (ἀπὸ τοῦ ἡλίου μετάστηθι)."
It is said that Alexander was so struck by this, and admired so much the
haughtiness and grandeur of the man who had nothing but scorn for him,
that he said to his followers, who were laughing and jesting about the philosopher as they went away,
"But truly, if I were not Alexander, I wish I were Diogenes.""