feat: rolled my own rusty SSG (#26)
* blank slate

* feat: init `Cargo` project

* feat(assets): add Edward Tufte's CSS and typefaces

* feat: add `justfile`

* feat(assets): add dithered `Alexander visits Diogenes in Corinth` pictures

ALEXANDER DIOGENEM CORINTH VISIT.

"Thereupon many statesmen and philosophers came to Alexander with their congratulations,
and he expected that Diogenes of Sinope also, who was tarrying in Corinth, would do likewise.
But since that philosopher took not the slightest notice of Alexander,
and continued to enjoy his leisure in the suburb Craneion, Alexander went in person to see him,
and he found him lying in the sun. Diogenes raised himself up a little when he saw
so many people coming towards him, and fixed his eyes upon Alexander.
And when that monarch addressed him with greetings, and asked if he wanted anything,
"Yes," said Diogenes, "stand a little out of my sun (ἀπὸ τοῦ ἡλίου μετάστηθι)."
It is said that Alexander was so struck by this, and admired so much the
haughtiness and grandeur of the man who had nothing but scorn for him,
that he said to his followers, who were laughing and jesting about the philosopher as they went away,
"But truly, if I were not Alexander, I wish I were Diogenes.""

* feat: SSG works

* feat(template): add `base.html` template

* feat(engine): render `LaTeX` expressions with `katex`

* chore(assets): move ET fonts to a folder

* feat(engine): highlight code blocks with `syntect`

* fix(engine): don't render `LaTeX` inside code blocks

* chore: bind dev server to all interfaces

* feat(engine): add sidenote and marginnote support

* fix: `tufte.css` overwrites, wrap all content with `<section>`

* feat(templates): add the banner template

* fix(engine): roll my own sidenote counter

* feat: add footer with quotes

* feat(engine,template): blog index

* fix(engine): allow subfolders in the blog directory

* feat(engine): move assets from blog subdir to build dir

* feat(blog): add Portable Self-Hosting

* feat(blog): add From Dice to Address

* chore(quotes): nits

* feat(engine): use Gruvbox Dark as the theme for code highlighting

* fix: don't link to `blog/*/index.html`, fix footer spacing, don't show footer on blog index

* feat(template): add sidebar for navigation

* chore: nits

* feat(engine): allow marking a post as draft

* chore: fix footer on mobile

* chore: add `.DS_Store` to .gitignore

* chore: stuff

* feat(engine): add ToC